[#13] - adding example page, increasing width and adding link to navbar [cf]
