+++ src/App.tsx
@@ -1,8 +1,9 @@
 import "./App.css";
 import { Home } from "./Pages/Home.tsx";
-import { Navbar } from "./components/Navbar.tsx";
+import { Navbar } from "./components/navBar/Navbar.tsx";
 import { Route, Routes } from "react-router-dom";
 import { Page2 } from "./Pages/Page2.tsx";
+import { Example } from "./Pages/Example.tsx";
 
 function App() {
   return (
@@ -11,6 +12,7 @@ function App() {
       <Routes>
         <Route path="/" element={<Home />} />
         <Route path="/page2" element={<Page2 />} />
+        <Route path="/example" element={<Example />} />
       </Routes>
     </>
   );
