+++ src/components/navBar/Navbar.tsx
@@ -10,6 +10,7 @@ export function Navbar() {
         </Typography>
         <Link to={"/"}>Home</Link>
         <Link to={"/page2"}>page2</Link>
+        <Link to={"/example"}>example</Link>
       </Toolbar>
     </AppBar>
   );
